// front/src/invoice/types/interfaces/Invoice.ts
import Item from "./Item";
import InvoiceStatus from "../types/Filter";

interface Invoice {
  senderAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  clientAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  id: string;
  createdAt: Date;
  paymentDue: Date;
  description: string;
  paymentTerms: number;
  clientName: string;
  clientEmail: string;
  status: InvoiceStatus;
  items: Item[];
  total: number;
}

export default Invoice;